move delegated stake interface to its own mod in staking primitives
